+++ CHANGES.md
@@ -1,3 +1,5 @@
+# Version 0.0.4
+
 # Version 0.0.3
 
 - Enable jump-to-definition in the comint buffer of the REPL